Adds support for a partial manifest file
A partial manifest file contains an ordered list of files, maybe sorted by run
time. This list is then combined with the glob, to make sure no files are missed.

The benefit here is that we can now specify our slowest specs in a manifest,
and let the rest go round-robin to get much better balancing.