Adding an option to add a git note to a successful build.
The namespace of the note is always "ci-<project_name>", and the note itself is configurable.  Repos can use this to add the "(ci=green)" note that hoist build expects.

At some point, we might want to allow projects to configure the namespace or to add a different note for each project within the repo, but for now this suffices to allow repos with a single project to add a note on successful build.