manually set build to succeeded to appease tests
because the build is now set to running immediately at build_part creation,
but not set to succeeded until after the build_state_update_job is run,
we end up with "running" instead of "succeeded". This commit fixes these tests
by (1) manually marking the build succeeded, and (2) changing the expectation.