Put therubyracer in the Gemfile
therubyracer is a JS runtime required by execjs, which is in turn required
by uglifier. therubyracer is not the fastest option but it is the most portable.

Doing this makes Kochiku easier to use by newcomers because they do not
need to install Node or another JS runtime on their machines.