Change the on_success_script from inside_copy to inside_repo
All uses of on_success_script so far have been to push additional
branches based on the contents of the commit. inside_repo has access to
the remote whereas inside_copy does not.

BREAKING CHANGE: on_success_script will no longer run with the working
directory checked out to the commit that was built. Instead, the script
is run from inside a git checkout of the repo but it should be treated
as bare repo. Avoid accessing files on the file system and instead rely
solely on git commands.