Switch cached repos inside build-partition to bare git repos
The use of full checkouts is a relic from the way things were done early
in Kochiku's life and is no longer necessary anymore. Bare repos are
cheaper and are easier to maintain. Another motivating factor is that
inside_copy can now reference origin/foo because it is downstream from a
bare repo instead of a checkout.

This undoes the earlier PR #199, on_success_script is once again being
executed via inside_copy instead of inside_repo.