Remove obsolete logic that was already commented out.
We decided against having a special rule for changes to kochiku.yml in the DependencyMap partitioner, as it is easy to inline rules for what tests should run in that case in the config section itself. This code was already commented out, with a ticket to re-enable it as some point. This commit just removes the commented-out code.