Include both versions of rename in net_files_changed_in_branch.
When looking for the complete list of files that have been changed on a branch, this should include files that have been renamed. Previously, only the new name was included. This change ensures that the orignal name of any renamed files is also returned by GitBlame#net_files_changed_in_branch.