fix: avoid dividing nil by 2 if the time manifest doesn't match files
After slicing the time manifest with available tests, you could end up with an empty set and therefore a `nil` minimum time. This ensures that the minimum time is at least 1.

This is a rare case that probably only arises when using the dependency_map partitioner, as it can limit the available tests to a small subset of all tests.